requirements.txt
python-telegram-bot==22.7
anthropic==2.30.0
openai>=1.0.0
requests>=2.28.0
python-docx>=1.0.0
openpyxl==3.1.5
